Added box and ldui64 opcodes

+++ src/main/java/com/verba/language/codegen/rendering/OpCodeRenderer.java
@@ -9,7 +9,7 @@ public interface OpCodeRenderer {
   void writeInt8(String label, int value);
   void writeInt16(String label, int value);
   void writeInt32(String label, int value);
-  void writeInt64(String label, int value);
+  void writeInt64(String label, long value);
 
   void writeString(String label, String value);
 }
